+++ eslint.config.js
@@ -1,6 +1,5 @@
 import js from '@eslint/js';
 import prettierConfig from 'eslint-config-prettier';
-import mocha from 'eslint-plugin-mocha';
 import importPlugin from 'eslint-plugin-import';
 
 /** @type {import('eslint').Linter.Config[]} */
@@ -16,14 +15,6 @@ export default [
     rules: {
       'import/order': 'error',
       ...prettierConfig.rules,
-    },
-  },
-  {
-    files: ['test/**/*.js'],
-    ...mocha.configs.flat.recommended,
-    rules: {
-      ...mocha.configs.flat.recommended.rules,
-      'mocha/no-setup-in-describe': 'off',
     },
   },
 ];