+++ vitest.setup.ts
@@ -1,6 +1,4 @@
 import { use } from 'chai';
 import chaiAsPromised from 'chai-as-promised';
-import sinonChai from 'sinon-chai';
 
 use(chaiAsPromised);
-use(sinonChai);